add types into config
