+++ src/packages/server/src/index.ts
@@ -1,6 +1,5 @@
 export * from '@apollo/server';
 export * from './plugins';
-export * from './types';
 export * from './config';
 
 export { default } from './server';
