--- src/packages/server/src/types.ts
@@ -1,18 +1,0 @@
-import { BaseContext } from '@apollo/server';
-
-export type MetadataHookParams<C> = {
-	context: C;
-	metadata?: { entities: any; enums: any };
-};
-export interface AdminMetadata {
-	enabled: boolean;
-	config?: any;
-	hooks?: {
-		beforeRead?: <C extends BaseContext>(
-			params: MetadataHookParams<C>
-		) => Promise<MetadataHookParams<C>>;
-		afterRead?: <C extends BaseContext>(
-			params: MetadataHookParams<C>
-		) => Promise<MetadataHookParams<C>>;
-	};
-}
